requirements.txt
aiobotocore==2.5.2
aiohttp==3.8.5
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
botocore==1.29.161
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
fsspec==2023.6.0
idna==3.4
jmespath==1.0.1
kafka-python==2.0.2
multidict==6.0.4
python-dateutil==2.8.2
requests==2.31.0
s3fs==2023.6.0
six==1.16.0
soupsieve==2.4.1
urllib3==1.26.16
wrapt==1.15.0
yarl==1.9.2
